fix(general-fix-fn): worked on general fix on the frontend - fixed an issue on seller login, that was giving an error due to recommendation products task 23 - changed the logic to generate the recommendation products only to buyers not to all logged in users - test coverage level up - re-designing the sellers dashboard for consistency design - resolving seller single product delete issue
[Fixes ]

+++ src/pages/SellerDashboard/viewItems.tsx
@@ -1,5 +1,5 @@
-import Sidebar from '@/components/SellerSidebar/sellerSidebar';
-import Header from '@/components/SellerDashHeader/sellerHeader';
+import Sidebar from '../../components/SellerSidebar/sellerSidebar';
+import AddProductForm from '../../components/productForm/productForm';
 import { useNavigate } from 'react-router-dom';
 import React, { useEffect, useState } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
@@ -7,14 +7,22 @@ import { RootState } from '../../redux/store/store';
 import { fetchProducts } from '../../redux/action/fetchProductAction';
 import { productRemove } from '../../redux/action/fetchProductAction';
 import ProductModal from '../../components/productViewModal/productModal';
+import moment from 'moment';
+import { faTrashAlt, faEye } from '@fortawesome/free-solid-svg-icons';
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 
 const ViewItems: React.FC = () => {
+  interface ProductToDelete {
+    productId: number;
+    name: string;
+  }
   const dispatch = useDispatch();
   const { data } = useSelector((state: RootState) => state.Allproducts.state);
   console.log(data);
   const itemsPerPage = 12;
   const [currentPage, setCurrentPage] = useState(1);
-  const [productDelete, setProductDelete] = useState({ productId: 0, name: '' });
+  const initialStateProductToDelete: ProductToDelete = { productId: 0, name: '' };
+  const [productDelete, setProductDelete] = useState<ProductToDelete>(initialStateProductToDelete);
   const [isDelete, setIsdelete] = useState(false);
   const navigate = useNavigate();
   useEffect(() => {
@@ -43,14 +51,15 @@ const ViewItems: React.FC = () => {
 
   function handleDelete(): void {
     const { productId } = productDelete;
-
-    dispatch(productRemove(productId) as any).then(({ payload }: any) => {
-      if (payload === 200) {
-        dispatch(fetchProducts() as any);
-        setProductDelete({ productId: 0, name: '' });
-        setIsdelete(false);
-      }
-    });
+    if (productId != 0) {
+      dispatch(productRemove(productId) as any).then(({ payload }: any) => {
+        if (payload === 200) {
+          dispatch(fetchProducts() as any);
+          setProductDelete({ productId: 0, name: '' });
+          setIsdelete(false);
+        }
+      });
+    }
   }
 
   function handleView(_id: string): void {
@@ -75,110 +84,150 @@ const ViewItems: React.FC = () => {
     }
     return buttons;
   };
-
+  const [visible, showForm] = useState(false);
+  const handleClose = () => showForm(false);
   return (
     <>
-      <div className="w-1/3">
-        <Sidebar />
-      </div>
-      <Header />
-      <label className="relative block">
-        <span className="sr-only">Search</span>
-
-        <input
-          className="placeholder:italic placeholder:text-slate-400 block bg-white w-48 border border-slate-300 rounded-md py-2 pl-9 pr-3 shadow-sm focus:outline-none focus:border-sky-500 focus:ring-sky-500 focus:ring-1 sm:text-sm"
-          placeholder="Search for product..."
-          type="text"
-          name="search"
-        />
-      </label>
-      <div className="w-2/3 ml-96 ">
-        <div className="flex justify-between items-center">
-          <h1 className="text-3xl font-bold mb-4 mt-4">In Stock Products</h1>
-          <input className="border border-slate-300 h-10 rounded pl-2" type="text" placeholder="Search product" onChange={(e) => handleSearch(e)} />
-        </div>
-
-        <div className="flex flex-col custom-md:grid custom-md:grid-cols-2 custom-md:gap-5 md:gap-7 lg:grid-cols-3">
-          {searchResult.length != 0
-            ? searchResult.map((product: any) => (
-                <div key={product.id} className="border border-slate-100 bg-white rounded-lg shadow-md p-4 mb-4h-[70vh] w-full mt-6">
-                  <h2 className="text-xl font-semibold mb-2">{product.name}</h2>
-                  <p className="text-gray-600 mb-2">{product.description}</p>
-                  <img src={product.image[0]} alt={product.name} className="w-1/2 mb-2 " />
-                  <p className="font-semibold mb-1">Price: {product.price}</p>
-                  <p className="font-semibold mb-1">Quantity: {product.quantity}</p>
-                  <div className="flex justify-between">
-                    <button type="button" data-modal-target="popup-modal" data-modal-toggle="popup-modal" className="bg-red-500 text-white font-semibold rounded px-4 py-2" onClick={() => handleDelete()}>
-                      Delete
+      <div className="h-screen">
+        <div className="flex h-screen">
+          <div className="w-1/6">
+            <Sidebar />
+          </div>
+          <div className="w-5/6 px-10 mt-10">
+            <div className="flex justify-between items-center mb-4">
+              <h1 className="text-3xl font-bold mb-4 mt-0 text-green-700">In Stock Products</h1>
+              <button onClick={() => showForm(true)} className="p-2 text-sm w-full md:w-64 shadow-lg bg-green-700 text-slate-50 hover:bg-green-900">
+                ADD NEW PRODUCT
+              </button>
+            </div>
+            <div className="flex justify-between items-center mb-4">
+              <h1 className="text-md">List of all products in your collection</h1>
+              <input className="border border-slate-300 h-10 w-full md:w-64 rounded pl-2" type="text" placeholder="Search product" onChange={(e) => handleSearch(e)} />
+            </div>
+            <div className="w-full pb-2 overflow-x-auto bg-white">
+              <table className="font-light w-full">
+                <thead className="bg-green-200">
+                  <tr>
+                    <th className="w-[2%] sm:w-5 text-start font-semibold px-3 py-4">ID</th>
+                    <th className="w-[20%] sm:w-32 text-start font-semibold px-3 py-4">Name</th>
+                    <th className="w-[40%] sm:w-64 text-start font-semibold px-3 py-4">Description</th>
+                    <th className="w-[16%] sm:w-20 text-start font-semibold px-3 py-4">Price</th>
+                    <th className="w-[20%] sm:w-24 text-start font-semibold px-3 py-4">Stock</th>
+                    <th className="w-[20%] sm:w-24 text-start font-semibold px-3 py-4">Category ID</th>
+                    <th className="w-[16%] sm:w-20 text-start font-semibold px-3 py-4">Quantity</th>
+                    <th className="w-[20%] sm:w-32 text-start font-semibold px-3 py-4">Expiration Date</th>
+                    <th className="w-[20%] sm:w-32 text-start font-semibold px-3 py-4">Action</th>
+                  </tr>
+                </thead>
+                <tbody>
+                  {searchResult.length != 0
+                    ? searchResult.map((product: any) => (
+                        <>
+                          <tr key={product.id} className="border-b-2 border-green-200">
+                            <td className="w-[2%] sm:w-5 text-start px-3 py-1">{product.id}</td>
+                            <td className="w-[20%] sm:w-24 text-start px-3 py-1">{product.name}</td>
+                            <td className="w-[40%] sm:w-64 text-start px-3 py-1">{product.description}</td>
+                            <td className="w-[16%] sm:w-20 text-start px-3 py-1">{product.price}</td>
+                            <td className="w-[20%] sm:w-24 text-start px-3 py-1">{product.stock}</td>
+                            <td className="w-[20%] sm:w-24 text-start px-3 py-1">{product.category_id}</td>
+                            <td className="w-[16%] sm:w-32 text-start px-3 py-1">{product.quantity}</td>
+                            <td className="w-[20%] sm:w-32 text-start px-3 py-1">{moment(product.expiredAt).format('MMM D,YYYY')}</td>
+                            <td className="w-[20%] sm:w-32 text-start px-6 py-3">
+                              <div className="flex justify-start">
+                                <button className="bg-green-700 text-white font-semibold rounded px-1 py-1 mr-2" onClick={() => handleView(product.id)}>
+                                  <FontAwesomeIcon icon={faEye} className="" />
+                                </button>
+                                <button
+                                  type="button"
+                                  data-modal-target="popup-modal"
+                                  data-modal-toggle="popup-modal"
+                                  className="bg-red-500 text-white font-semibold rounded px-1 py-1"
+                                  // onClick={() => handleDelete()}
+                                  onClick={() => {
+                                    setProductDelete({ productId: product.id, name: product.name });
+                                    setIsdelete(true);
+                                  }}
+                                >
+                                  <FontAwesomeIcon icon={faTrashAlt} className="" />
+                                </button>
+                              </div>
+                            </td>
+                          </tr>
+                        </>
+                      ))
+                    : currentProducts.map((product: any) => (
+                        <tr key={product.id} className="border-b-2 border-green-200">
+                          <td className="w-[2%] sm:w-5 text-start px-3 py-1">{product.id}</td>
+                          <td className="w-[20%] sm:w-24 text-start px-3 py-1">{product.name}</td>
+                          <td className="w-[40%] sm:w-64 text-start px-3 py-1">{product.description}</td>
+                          <td className="w-[16%] sm:w-20 text-start px-3 py-1">{product.price}</td>
+                          <td className="w-[20%] sm:w-24 text-start px-3 py-1">{product.stock}</td>
+                          <td className="w-[20%] sm:w-24 text-start px-3 py-1">{product.category_id}</td>
+                          <td className="w-[16%] sm:w-32 text-start px-3 py-1">{product.quantity}</td>
+                          <td className="w-[20%] sm:w-32 text-start px-3 py-1">{moment(product.expiredAt).format('MMM D,YYYY')}</td>
+                          <td className="w-[20%] sm:w-32 text-start px-6 py-3">
+                            <div className="flex justify-start">
+                              <button className="bg-green-700 text-white font-semibold rounded px-1 py-1 mr-2" onClick={() => handleView(product.id)}>
+                                <FontAwesomeIcon icon={faEye} className="" />
+                              </button>
+                              <button
+                                type="button"
+                                data-modal-target="popup-modal"
+                                data-modal-toggle="popup-modal"
+                                className="bg-red-500 text-white font-semibold rounded px-1 py-1"
+                                onClick={() => {
+                                  setProductDelete({ productId: product.id, name: product.name });
+                                  setIsdelete(true);
+                                }}
+                              >
+                                <FontAwesomeIcon icon={faTrashAlt} className="" />
+                              </button>
+                            </div>
+                          </td>
+                        </tr>
+                      ))}
+                </tbody>
+              </table>
+              <AddProductForm onClose={handleClose} visible={visible} />
+            </div>
+            <div className="flex justify-center">
+              {totalPages > 1 && (
+                <div className="w-[95%] m-auto flex justify-end mt-4">
+                  <button disabled={currentPage === 1} onClick={() => goToPage(currentPage - 1)} className="text-sm font-semibold mr-3 text-[#003D29]">
+                    Previous
+                  </button>
+                  {renderNumberButtons()}
+                  <button disabled={currentPage === totalPages} onClick={() => goToPage(currentPage + 1)} className="text-sm font-semibold ml-1 text-[#003D29]">
+                    Next
+                  </button>
+                </div>
+              )}
+            </div>
+            {currentProductId && <ProductModal productId={currentProductId} onClose={() => setCurrentProductId(null)} />}
+            {isDelete && (
+              <div className="fixed inset-0 flex items-center justify-center z-50">
+                <div className="fixed inset-0 bg-black opacity-50"></div>
+                <div className="relative bg-white w-1/3 rounded-lg shadow-lg">
+                  <div className="p-4">
+                    <h2 className="text-xl text-red-500 font-semibold mb-5">Warning!</h2>
+                    <p>Are you sure you want to delete this product?</p>
+                    <p><span className='font-semibold'>Product ID: </span>{productDelete.productId}, </p>
+                    <p><span className='font-semibold'>Product Name: </span>{productDelete.name} </p>
+                  </div>
+                  <div className="flex justify-between p-4">
+                    <button className="border border-gray-400 hover:bg-gray-400 text-gray-800 py-2 px-4 rounded" onClick={() => setIsdelete(false)}>
+                      Close
                     </button>
-                    <button className="bg-green-700 text-white font-semibold rounded px-4 py-2" onClick={() => handleView(product.id)}>
-                      View
+                    <button className="bg-red-500 text-white font-semibold rounded px-4 py-2" onClick={() => handleDelete()}>
+                      Delete
                     </button>
                   </div>
                 </div>
-              ))
-            : currentProducts.map((product: any) => (
-                <div key={product.id} className="border border-slate-100 bg-white rounded-lg shadow-md p-4 mb-4h-[70vh] w-full mt-6">
-                  <h2 className="text-xl font-semibold mb-2">{product.name}</h2>
-                  <p className="text-gray-600 mb-2">{product.description}</p>
-                  <img src={product.image[0]} alt={product.name} className="w-1/2 mb-2 " />
-                  <p className="font-semibold mb-1">Price: {product.price}</p>
-                  <p className="font-semibold mb-1">Quantity: {product.quantity}</p>
-                  <div className="flex justify-between">
-                    <button
-                      type="button"
-                      data-modal-target="popup-modal"
-                      data-modal-toggle="popup-modal"
-                      className="bg-red-500 text-white font-semibold rounded px-4 py-2"
-                      onClick={() => {
-                        setProductDelete({ productId: product.id, name: product.name });
-                        setIsdelete(true);
-                      }}
-                    >
-                      Delete
-                    </button>
-                    <button className="bg-green-700 text-white font-semibold rounded px-4 py-2" onClick={() => handleView(product.id)}>
-                      View
-                    </button>
-                  </div>
-                </div>
-              ))}
-        </div>
-
-        <div className="flex justify-center">
-          {totalPages > 1 && (
-            <div className="w-[95%] m-auto flex justify-end mt-4">
-              <button disabled={currentPage === 1} onClick={() => goToPage(currentPage - 1)} className="text-sm font-semibold mr-3 text-[#003D29]">
-                Previous
-              </button>
-              {renderNumberButtons()}
-              <button disabled={currentPage === totalPages} onClick={() => goToPage(currentPage + 1)} className="text-sm font-semibold ml-1 text-[#003D29]">
-                Next
-              </button>
-            </div>
-          )}
-        </div>
-      </div>
-      {currentProductId && <ProductModal productId={currentProductId} onClose={() => setCurrentProductId(null)} />}
-      {isDelete && (
-        <div className="fixed inset-0 flex items-center justify-center z-50">
-          <div className="fixed inset-0 bg-black opacity-50"></div>
-          <div className="relative bg-white w-1/3 rounded-lg shadow-lg">
-            <div className="p-4">
-              <h2 className="text-xl text-red-500 font-semibold mb-5">Warning!</h2>
-              <p>Are you sure you want to delete {productDelete.name} product?</p>
-            </div>
-            <div className="flex justify-between p-4">
-              <button className="border border-gray-400 hover:bg-gray-400 text-gray-800 py-2 px-4 rounded" onClick={() => setIsdelete(false)}>
-                Close
-              </button>
-              <button className="bg-red-500 text-white font-semibold rounded px-4 py-2" onClick={() => handleDelete()}>
-                Delete
-              </button>
-            </div>
+              </div>
+            )}
           </div>
         </div>
-      )}
+      </div>
     </>
   );
 };
